backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2
boto3==1.34.8

# ML dependencies
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.4
scikit-learn==1.3.2
transformers==4.36.2
diffusers==0.25.0
accelerate==0.25.0
safetensors==0.4.1
rembg==2.0.50
mediapipe==0.10.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0